Raylib test program
I did a sample raylib program in Clion to know if i can actually understand and do it. I did most of it myself. having a variable used in multiple files is very hard to maintain. So far so good. need to learn more from learncpp to have better structure for my code.